Fixed the job status diagram
Now, we use the on_import_error_found method to change the status of the jobs. Thus, the first time we encounter an error, the status is set to "auth_required". If later, the job is executed again, and still finds an error, the status of the job will change to "data_unreachable"