Using token authentication for wikifactory exporter
Fixed the error by which the export_manifest method was not returning an error object. It was returning None.
The Auth token is used in the wikifactory_api_request method in the headers
Fixed the asserts of the associated tests, since we are now returning an error object instead of None if something was wrong during the export process