Updated model to use pydantic
Removed unused file thing, which has been replaced by the Element class
Additionally, we have added the BACKEND_URL env var to the docker-compose, which is used to create the celery instance. In particular, it is required to serve the /job manifest operation. In order to be able to wait for the manifest generation, we need to specify the backend url