Created populate_project_description method
That method is defined in base.py, and hence inherited by all the importers.

The git importer scans the root level folder searching for the readme file
The google drive importer uses the "description" field of the root_item. We could later implement a fill policy, in case this field is empty (e.g. fill the creator, creation date and so on).